+++ src/db/migrations/migrations.js
@@ -3,12 +3,14 @@
 import journal from './meta/_journal.json';
 import m0000 from './0000_lists.sql';
 import m0001 from './0001_items.sql';
+import m0002 from './0002_shops.sql';
 
   export default {
     journal,
     migrations: {
       m0000,
-m0001
+m0001,
+m0002
     }
   }